+++ src/test/java/blackjack/domain/card/HandTest.java
@@ -1,8 +1,8 @@
 package blackjack.domain.card;
 
-import static blackjack.domain.Result.DRAW;
-import static blackjack.domain.Result.LOSE;
-import static blackjack.domain.Result.WIN;
+import static blackjack.domain.result.Result.DRAW;
+import static blackjack.domain.result.Result.LOSE;
+import static blackjack.domain.result.Result.WIN;
 import static blackjack.domain.card.Denomination.ACE;
 import static blackjack.domain.card.Denomination.J;
 import static blackjack.domain.card.Denomination.K;
