+++ src/test/java/blackjack/domain/participant/DealerTest.java
@@ -9,7 +9,7 @@ import static org.assertj.core.api.Assertions.assertThatIllegalStateException;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.junit.jupiter.api.Assertions.assertAll;
 
-import blackjack.domain.Money;
+import blackjack.domain.money.Money;
 import blackjack.domain.card.Card;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
